Modify Word Card Component HTML and CSS
This commit adds a class to the delete word button in wordlist-card.component.html. This class is then manipulated in the CSS to customize the margins, alignment, and the background color for when the button is hovered over.

Co-authored-by:
Jeffrey Miller<mill7707@morris.umn.edu>
Jadyn Sondrol <Sondr037@morris.umn.edu> Natasha Zebrev <zebre001@morris.umn.edu>